requirements.txt
kafka-python
python-dotenv
google-generativeai
